adjust the translation documentation, and remove custom_locales

+++ config/deploy.rb
@@ -9,7 +9,7 @@ set :stages, %w(prod)
 set :linked_files, %w{config/database.yml config/secrets.yml config/newrelic.yml public/robots.txt}
 
 # Default value for linked_dirs is []
-set :linked_dirs, %w{bin log tmp/pids tmp/cache tmp/sockets vendor/bundle public/system public/sitemaps config/custom_locales}
+set :linked_dirs, %w{bin log tmp/pids tmp/cache tmp/sockets vendor/bundle public/system public/sitemaps}
 
 before 'deploy', 'sidekiq:stop'
 after 'deploy:publishing', 'deploy:restart'
@@ -25,6 +25,7 @@ namespace :deploy do
       within release_path do
         with :rails_env => fetch(:rails_env) do
           execute :rake, "import_translations_from_yml"
+          execute :rake, "write_tolk_to_disk"
         end
       end
     end
@@ -34,18 +35,19 @@ after 'deploy:published', 'deploy:install_translations'
 
 namespace :translation do
   task :download do
+    local_diff = `git status --untracked-files=no --porcelain`
     on primary(:app) do
-      download! "#{release_path}/config/custom_locales/", "config/", recursive: true
-      FileUtils.rm_rf "config/locales"
-      FileUtils.mv "config/custom_locales/", "config/locales"
+      if local_diff.empty?
+        FileUtils.rm_rf "config/locales"
+        download! "#{release_path}/config/locales/", "config/", recursive: true
+      else
+        puts "****** ERROR *******"
+        puts "For safety purposes you cannot run this with a dirty local git directory"
+        puts "Changes local:"
+        puts "#{local_diff}"
+      end
     end
   end
-
-  task :prime do
-    on primary(:app) do
-       execute "cp -R #{release_path}/config/locales/* #{release_path}/config/custom_locales/"
-     end
-   end
 end
 
 set :whenever_command,      ->{ [:bundle, :exec, :whenever] }
